# federation-agricultural/src/main/java/mg/hei/federation_agricole/model/enums/Bank.java
package mg.hei.federation_agricole.model.enums;

public enum Bank {
    BRED
        , MCB
        , BMOI
        , BOA
        , BGFI
        , AFG
        ,ACCES_BAQUE
        ,BAOBAB
        , SIPEM
}